% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{graph_relation}
\alias{graph_relation}
\title{Scatterplot for two Numeric Variables}
\usage{
graph_relation(x, y)
}
\arguments{
\item{x}{A numeric vector, I name it x because it represents the x-axis values.}

\item{y}{A numeric vector that must have the same length with x. I name it y because it represents the y-axis values}
}
\value{
A graph "ggplot", a scatter plot of the two variables
}
\description{
This function creates a scatterplot to demonstrate the relationship between two numeric variables.
}
\examples{
#Create a scatter plot for two variables x and y
x <- c(43,23,78,100,88,65,49)
y <- c(19,20,34,54,33,24,12)
graph_relation(x,y)
#Create a scatter plot for two variables a and b
a <- rnorm(20,mean=5, sd=1)
b <- rnorm(20, mean=10, sd=2)
graph_relation(a,b)
}
